+++ app/Http/Controllers/ProfileController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\admin;
 use App\Models\instruktur;
 use App\Models\kelas;
+use App\Models\nilai;
+use App\Models\nilai_kuis;
 use App\Models\siswa;
 use App\Models\users;
 use Illuminate\Http\Request;
@@ -22,6 +24,14 @@ class ProfileController extends Controller
         $bidang = $request->bidang_siswa;
         $tglMasuk = $request->tglMasuk;
         $status = $request->status_siswa;
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+
+
         if (trim($nama) === '' || trim($email) === ''|| trim($nohp) === ''|| trim($alamat) === '') {
             return back()->with(["error_add" => "Input tidak boleh kosong atau hanya berisi spasi!"]);
         }
@@ -31,9 +41,9 @@ class ProfileController extends Controller
 
         // cek kuota kelas
         $cekKelas = kelas::where("id_kelas",'=',$kelas)->first();
-        $cekjmlsiswa = siswa::where("id_kelas",'=',$kelas)->count();
-        // dd($cekjmlsiswa);
-        if($cekjmlsiswa == $cekKelas->kuota){
+        $cekjmlsiswa = siswa::where("id_kelas",'=',$kelas)->whereIn('status', ['aktif', 'mangkir'])->count();
+        // dd($cekKelas);
+        if($cekjmlsiswa == $cekKelas->kuota_siswa){
             return back()->with(['error_add' => 'Kelas sudah penuh!']);
 
         }
@@ -60,11 +70,17 @@ class ProfileController extends Controller
         }
 
         // ==== menentukan id siswa ====
-        $cekCountSiswa = siswa::count();
+        $lastSiswaId = siswa::select(DB::raw('CAST(SUBSTRING(id_siswa, 6) AS UNSIGNED) AS angka'))
+            ->orderBy('angka', 'desc')
+            ->first();
+
+        // Ambil hanya angka dari ID terakhir (jika ada)
+        $lastNumber = $lastSiswaId ? $lastSiswaId->angka : 0;
+        //dd($lastNumber);
 
         //dd($cekCountSiswa);
         $insertData = DB::table("siswas")->insert([
-            "id_siswa" => "Siswa".$cekCountSiswa + 2,
+            "id_siswa" => "Siswa".$lastNumber + 1,
             "nama" => $nama,
             "email" => $email,
             "no_hp" => $nohp,
@@ -88,25 +104,29 @@ class ProfileController extends Controller
             return abort(403);
 
         }
+
+        //cek siswa udah ada data nilai apa belom
+        $cekNilai = nilai_kuis::where("id_siswa",'=',$id)->first();
+        $cekNilai2 = nilai::where("id_siswa",'=',$id)->first();
+        if($cekNilai == true || $cekNilai2 == true){
+            return back()->with(['error_delete' => "Siswa telah memiliki data nilai!"]);
+        }
         // dd($id);
         // getEmail yang dipake siswa sesuai id siswa
         $getEmail = DB::table("siswas")->select("email")->where('id_siswa','=', $id)->first();
+        //dd($getEmail);
         $deleteSiswa = DB::table('siswas')->where('id_siswa','=', $id)->delete();
         if($deleteSiswa == true){
-            //  menghapus data siswa tpi emailnya udah kedaftar di table users 
-            $getAkun = DB::table('users')->where('email','=', $getEmail->email)->first();
-            // klo akun nya ktmu di table users brarti email yagn ada di table users juga dihapus
-            if($getAkun == true){
-                $deleteEmail = DB::table('users')->where('email','=', $getEmail->email)->delete();
-                if($deleteEmail == true){
+            if($getEmail == true){
+                $deleteAkun = DB::table('users')->where('email','=', $getEmail->email)->delete();
+                if($deleteAkun == true){
                     return back()->with(['sukses_delete' => "Siswa berhasil dihapus!"]);
+
                 }else{
                     return back()->with(['error_delete' => "Akun siswa gagal dihapus!"]);
                 }
-            }else{
-                // ini ketika delete siswa nya berhasil tpi email blm kedaftar
-                return back()->with(['sukses_delete' => "Siswa berhasil dihapus!"]);
             }
+            return back()->with(['sukses_delete' => "Siswa berhasil dihapus!"]);
             
         }else{
             return back()->with(['error_delete' => "Siswa gagal dihapus!"]);
@@ -120,9 +140,23 @@ class ProfileController extends Controller
         $tglLulus = $request->tglLulus;
         $no_hp = $request->nohpEdit;
         $alamat = $request->alamatEdit;
+        $tglMasuk = $request->tglMasukEdit;
         $id_bidang = $request->bidang_siswaEdit;
         $id_kelas = $request->kelas_siswaEdit;
         $status = $request->status_siswaEdit;
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+        if($tglLulus != null){
+            $tahunLulus = (int) date('Y', strtotime($tglLulus));
+            if ($tglLulus <= $tglMasuk || $tahunLulus > $currentYear) {
+                return back()->with(["error_add" => "Format tahun lulus tidak valid!"]);
+            }
+        }
+
         if (trim($nama) === '' || trim($email) === ''|| trim($no_hp) === ''|| trim($alamat) === '') {
             return back()->with(["error_add" => "Input tidak boleh kosong atau hanya berisi spasi!"]);
         }
@@ -130,11 +164,12 @@ class ProfileController extends Controller
             return back()->with(['error_add' => 'Format nomor telepon tidak valid!']);
         }
         $cekKelasSiswa = siswa::where('id_siswa','=',$id)->first();
-        if($id_kelas !== $cekKelasSiswa->id_kelas){
+        if($id_kelas != $cekKelasSiswa->id_kelas){
+            //dd($id_kelas);
             $cekKelas = kelas::where("id_kelas",'=',$id_kelas)->first();
-            $cekjmlsiswa = siswa::where("id_kelas",'=',$id_kelas)->count();
+            $cekjmlsiswa = siswa::where("id_kelas",'=',$id_kelas)->whereIn('status', ['aktif', 'mangkir'])->count();
             // dd($cekjmlsiswa);
-            if($cekjmlsiswa == $cekKelas->kuota){
+            if($cekjmlsiswa == $cekKelas->kuota_siswa){
                 return back()->with(['error_add' => 'Kelas sudah penuh!']);
     
             }
@@ -172,22 +207,35 @@ class ProfileController extends Controller
                 "no_hp" => $no_hp,
                 "alamat" => $alamat,
                 "id_bidang" => $id_bidang,
+                "tgl_masuk" => $tglMasuk,
                 "id_kelas" => $id_kelas,
                 "status" => $status,
             ]);
     
             if($updateSiswa == true){
                 // jika data email yang dimasukkan adlah email baru maka akan di edit, jika sama , brarti tidak di edit
-    
-      
                 if($getEmailSiswa->email != $email){
                     // kondis  jika email yang dimasukkan berubah atau email baru
                     $updateEmail = DB::table("users")->where("email",'=',$getEmailSiswa->email)->update([
                         "email" => $email
                     ]);
-    
-                    if($updateEmail == true){
-                        return back()->with(["sukses_edit" => "Data siswa berhasil diubah!"]);
+                }
+                // cek ketika status yang diubah tuh selain aktif
+                if($status != "aktif"){
+                    // mengubah status pada akun
+                    // cek dlu dia udah register apa blom
+                    $cekAkun = users::where("email",'=',$email)->first();
+                    if($cekAkun == true){
+                        DB::table("users")->where("email",'=',$email)->update([
+                            "status_akun" => "tidak_aktif"
+                        ]);
+                    }
+                }else{
+                    $cekAkun = users::where("email",'=',$email)->first();
+                    if($cekAkun == true){
+                        DB::table("users")->where("email",'=',$email)->update([
+                            "status_akun" => "aktif"
+                        ]);
                     }
                 }
                 return back()->with(["sukses_edit" => "Data siswa berhasil diubah!"]);
@@ -215,9 +263,22 @@ class ProfileController extends Controller
                     $updateEmail = DB::table("users")->where("email",'=',$getEmailSiswa->email)->update([
                         "email" => $email
                     ]);
-    
-                    if($updateEmail == true){
-                        return back()->with(["sukses_edit" => "Data siswa berhasil diubah!"]);
+                }
+                if($status != "aktif"){
+                    // mengubah status pada akun
+                    // cek dlu dia udah register apa blom
+                    $cekAkun = users::where("email",'=',$email)->first();
+                    if($cekAkun == true){
+                        DB::table("users")->where("email",'=',$email)->update([
+                            "status_akun" => "tidak_aktif"
+                        ]);
+                    }
+                }else{
+                    $cekAkun = users::where("email",'=',$email)->first();
+                    if($cekAkun == true){
+                        DB::table("users")->where("email",'=',$email)->update([
+                            "status_akun" => "aktif"
+                        ]);
                     }
                 }
                 return back()->with(["sukses_edit" => "Data siswa berhasil diubah!"]);
@@ -234,6 +295,15 @@ class ProfileController extends Controller
         $nohp = $request->nohp;
         $alamat = $request->alamat;
         $tglMasuk = $request->tglMasuk;
+
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+
+
         if (trim($nama) === '' || trim($email) === ''|| trim($nohp) === ''|| trim($alamat) === '') {
             return back()->with(["error_add" => "Input tidak boleh kosong atau hanya berisi spasi"]);
         }
@@ -261,10 +331,16 @@ class ProfileController extends Controller
         }
 
         // ==== menentukan id siswa ====
-        $cekCountIns = instruktur::count();
+        //$cekCountIns = instruktur::count();
 
+        $lastInsId = instruktur::select(DB::raw('CAST(SUBSTRING(id_ins, 6) AS UNSIGNED) AS angka'))
+            ->orderBy('angka', 'desc')
+            ->first();
+
+        // Ambil hanya angka dari ID terakhir (jika ada)
+        $lastNumber = $lastInsId ? $lastInsId->angka : 0;
         $insertData = DB::table("instrukturs")->insert([
-            "id_ins" => "Ins".$cekCountIns + 1,
+            "id_ins" => "Ins".$lastNumber + 1,
             "nama_ins" => $nama,
             "email_ins" => $email,
             "no_hp_ins" => $nohp,
@@ -285,7 +361,12 @@ class ProfileController extends Controller
             return abort(403);
 
         }
-        // dd($id);
+        // cek instruktur lgi ngajar di kelas apa ga
+        $cekKelas = kelas::where("id_ins",'=',$id)->first();
+        if($cekKelas == true){
+            return back()->with(['error_delete' => "Instruktur sedang terhubung dengna kelas!"]);
+            
+        }
         // getEmail yang dipake instruktur sesuai id instruktur
         $getEmail = DB::table("instrukturs")->select("email_ins")->where('id_ins','=', $id)->first();
         $deleteSiswa = DB::table('instrukturs')->where('id_ins','=', $id)->delete();
@@ -313,8 +394,18 @@ class ProfileController extends Controller
         $id = $request->instrukturs;
         $nama = $request->namaEdit;
         $email = $request->emailEdit;
+        $tglMasuk = $request->tglMasukEdit;
         $no_hp = $request->nohpEdit;
         $alamat = $request->alamatEdit;
+
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+
+
         if (!preg_match('/^(\+?[0-9]{1,3}[0-9]{1,}|[0-9]{1,})$/', $request->nohpEdit)) {
             return back()->with(['error_add' => 'Format nomor telepon tidak valid!']);
         }
@@ -350,6 +441,7 @@ class ProfileController extends Controller
             "nama_ins" => $nama,
             "email_ins" => $email,
             "no_hp_ins" => $no_hp,
+            "tgl_masuk_ins" => $tglMasuk,
             "alamat_ins" => $alamat,
         ]);
 
@@ -387,6 +479,14 @@ class ProfileController extends Controller
         $nohp = $request->nohp;
         $alamat = $request->alamat;
         $tglMasuk = $request->tglMasuk;
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+
+
         if (trim($nama) === '' || trim($email) === ''|| trim($nohp) === ''|| trim($alamat) === '') {
             return back()->with(["error_add" => "Input tidak boleh kosong atau hanya berisi spasi!"]);
         }
@@ -429,6 +529,15 @@ class ProfileController extends Controller
         $nama = $request->namaEdit;
         $email = $request->emailEdit;
         $nohp = $request->nohpEdit;
+        $tglMasuk = $request->tglMasukEdit;
+
+        $currentYear = (int) date('Y'); // Tahun saat ini
+        $tahunMasuk = (int) date('Y', strtotime($tglMasuk));
+        if ($tahunMasuk < 2023 || $tahunMasuk > $currentYear) {
+            return back()->with(["error_add" => "Format tahun pada tanggal masuk tidak valid!"]);
+        }
+
+
         $alamat = $request->alamatEdit;
         if (!preg_match('/^\+?[0-9\s]{8,}$/', $request->nohpEdit)) {
             return back()->with(['error_add' => 'Format nomor HP tidak valid.']);
@@ -438,14 +547,15 @@ class ProfileController extends Controller
         }
         $getEmailAdmin = DB::table("admins")->where("id_adm",'=',$admins)->first();
 
-        $cekEmail1 = users::select("email")->from('users')->where("email",'=',$email)->where("id_akun",'=',$getEmailAdmin->id_akun)->first();
+        $cekEmail1 = users::select("email")->from('users')->where("email",'=',$email)->where("id_akun",'!=',$getEmailAdmin->id_akun)->first();
         $cekEmail = instruktur::select("email_ins")->from('instrukturs')->where("email_ins",'=',$email)->first();
         $ceknohp = instruktur::select("no_hp_ins")->from('instrukturs')->where("no_hp_ins",'=',$nohp)->first();
         $cekEmailSiswa = siswa::select("email")->from('siswas')->where("email",'=',$email)->first();
         $ceknohpSiswa = siswa::select("no_hp")->from('siswas')->where("no_hp",'=',$nohp)->first();
         $cekEmailadmin = admin::select("email_adm")->from('admins')->where("email_adm",'=',$email)->where("id_adm",'!=',$admins)->first();
         $ceknohpadmin = admin::select("no_hp_adm")->from('admins')->where("no_hp_adm",'=',$nohp)->where("id_adm",'!=',$admins)->first();
-
+        //dd($cekEmail1);
+        
 
 
         
@@ -465,6 +575,7 @@ class ProfileController extends Controller
             "nama_adm" => $nama,
             "email_adm" => $email,
             "no_hp_adm" => $nohp,
+            "tgl_masuk_adm" => $tglMasuk,
             "alamat_adm" => $alamat,
         ]);
         if($updateAdm == true){
